# APItest/SwaggerUI api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SK API",
    "description": "SK智慧校园API文档",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v114.514"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/sms/gradeController/saveOrUpdateGrade": {
      "post": {
        "tags": [
          "Grade"
        ],
        "summary": "增加或修改年级",
        "description": "增加或者修改年级信息",
        "operationId": "saveOrUpdateGrade",
        "parameters": [
          {
            "name": "grade",
            "in": "query",
            "description": "JSON格式的年级信息",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Grade"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/sms/clazzController/saveOrUpdateClazz": {
      "post": {
        "tags": [
          "Clazz"
        ],
        "summary": "增加或者修改班级",
        "description": "增加或者修改班级信息",
        "operationId": "saveOrUpdateClazz",
        "parameters": [
          {
            "name": "clazz",
            "in": "query",
            "description": "JSON格式的班级信息",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Clazz"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/sms/adminController/saveOrUpdateAdmin": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "增加或修改管理员",
        "description": "增加或修改管理员信息",
        "operationId": "saveOrUpdateAdmin",
        "parameters": [
          {
            "name": "admin",
            "in": "query",
            "description": "JSON格式的Admin对象",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Admin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/gradeController/getGrades/{pageNo}/{pageSize}": {
      "get": {
        "tags": [
          "Grade"
        ],
        "summary": "分页查年级",
        "description": "分页带条件查询年级信息",
        "operationId": "getGrades_1",
        "parameters": [
          {
            "name": "pageNo",
            "in": "path",
            "description": "当前页码",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "path",
            "description": "每页显示条数",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "gradeName",
            "in": "query",
            "description": "年级名称",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/clazzController/getClazzsByOpr/{pageNo}/{pageSize}": {
      "get": {
        "tags": [
          "Clazz"
        ],
        "summary": "分页查班级",
        "description": "分页带条件查询班级信息",
        "operationId": "getClazzByOpr",
        "parameters": [
          {
            "name": "pageNo",
            "in": "path",
            "description": "当前页码",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "path",
            "description": "每页显示条数",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "clazz",
            "in": "query",
            "description": "班级信息",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Clazz"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/sms/adminController/getAllAdmin/{pageNo}/{pageSize}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "分页查管理员",
        "description": "分页带条件查询管理员信息",
        "operationId": "getAllAdmin",
        "parameters": [
          {
            "name": "pageNo",
            "in": "path",
            "description": "页码数",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "path",
            "description": "页大小",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "adminName",
            "in": "query",
            "description": "管理员名字",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/gradeController/deleteGrade": {
      "delete": {
        "tags": [
          "Grade"
        ],
        "summary": "删除年级",
        "description": "删除单个和多个年级",
        "operationId": "deleteGrade",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "要删除的多个年级的ID的JSON数组",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/sms/clazzController/deleteClazz": {
      "delete": {
        "tags": [
          "Clazz"
        ],
        "summary": "删除班级",
        "description": "删除单个和多个班级",
        "operationId": "deleteClazz",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "要删除的多个班级的ID的JSON数组",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/sms/adminController/deleteAdmin": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "删除管理员",
        "description": "删除单个或者多个管理员信息",
        "operationId": "deleteAdmin",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "管理员id数组",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ResultObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "数据",
            "example": null
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "Grade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "年级Id",
            "format": "int32",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "年级名称",
            "example": "软件工程"
          },
          "introducation": {
            "type": "string",
            "description": "年级介绍",
            "example": "软件工程"
          },
          "manager": {
            "type": "string",
            "description": "年级主任姓名",
            "example": "张三"
          },
          "email": {
            "type": "string",
            "description": "年级主任邮箱",
            "example": "2421361985"
          },
          "telephone": {
            "type": "string",
            "description": "年级主任电话",
            "example": "18888888888"
          }
        }
      },
      "Clazz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "班级Id",
            "format": "int32",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "班级名称",
            "example": "软件工程1班"
          },
          "number": {
            "type": "string",
            "description": "班级人数",
            "example": "50"
          },
          "introducation": {
            "type": "string",
            "description": "班级介绍",
            "example": "软件工程1班"
          },
          "headmaster": {
            "type": "string",
            "description": "班主任姓名",
            "example": "张三"
          },
          "telephone": {
            "type": "string",
            "description": "班主任电话",
            "example": "18888888888"
          },
          "email": {
            "type": "string",
            "description": "班主任邮箱",
            "example": "2421361985"
          },
          "gradeName": {
            "type": "string",
            "description": "班级所属年级",
            "example": "软件工程"
          }
        }
      },
      "Admin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "管理员id",
            "format": "int32",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "姓名",
            "example": "张三"
          },
          "gender": {
            "type": "string",
            "description": "性别",
            "example": "男"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "123456"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "2425398986@qq.com"
          },
          "telephone": {
            "type": "string",
            "description": "电话",
            "example": "18888888888"
          },
          "address": {
            "type": "string",
            "description": "地址",
            "example": "广东省广州市天河区"
          },
          "portraitPath": {
            "type": "string",
            "description": "头像的图片路径",
            "example": "http://localhost:8080/img/1.jpg"
          }
        }
      },
...
    }
  }
}